MC-5-3: refactor arcitechure by FSD rules, IN-progress feature Modules
