+++ src/entities/Module/types.ts
@@ -1,0 +1,8 @@
+import { WordEntity } from 'src/features/words/types/word';
+
+export interface IModule {
+  id: string;
+  title: string;
+  description: string;
+  words: WordEntity[];
+}
